refactor: auth에서 member 의존성 제거
