+++ src/main/java/nextstep/auth/User.java
@@ -1,0 +1,13 @@
+package nextstep.auth;
+
+public interface User {
+    boolean checkPassword(String password);
+
+    Long getId();
+
+    String getEmail();
+
+    Integer getAge();
+
+    String getPassword();
+}
