+++ src/main/java/nextstep/auth/authentication/AuthenticationInterceptor.java
@@ -1,8 +1,9 @@
 package nextstep.auth.authentication;
 
+import nextstep.auth.User;
+import nextstep.auth.authentication.after.AfterAuthentication;
+import nextstep.auth.authentication.converter.AuthenticationConverter;
 import nextstep.auth.context.Authentication;
-import nextstep.member.application.CustomUserDetailsService;
-import nextstep.member.domain.LoginMember;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
@@ -13,10 +14,10 @@ import java.io.IOException;
 
 public class AuthenticationInterceptor implements HandlerInterceptor {
     private final AuthenticationConverter authenticationConverter;
-    private final CustomUserDetailsService userDetailsService;
+    private final UserDetailsService userDetailsService;
     private final AfterAuthentication afterAuthentication;
 
-    public AuthenticationInterceptor(AuthenticationConverter authenticationConverter, CustomUserDetailsService userDetailsService, AfterAuthentication afterAuthentication) {
+    public AuthenticationInterceptor(AuthenticationConverter authenticationConverter, UserDetailsService userDetailsService, AfterAuthentication afterAuthentication) {
         this.authenticationConverter = authenticationConverter;
         this.userDetailsService = userDetailsService;
         this.afterAuthentication = afterAuthentication;
@@ -33,13 +34,13 @@ public class AuthenticationInterceptor implements HandlerInterceptor {
 
     public Authentication authenticate(AuthenticationToken token) {
         String principal = token.getPrincipal();
-        LoginMember userDetails = userDetailsService.loadUserByUsername(principal);
+        User userDetails = userDetailsService.loadUserByUsername(principal);
         checkAuthentication(userDetails, token);
 
         return new Authentication(userDetails);
     }
 
-    public void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
+    public void checkAuthentication(User userDetails, AuthenticationToken token) {
         if (userDetails == null) {
             throw new AuthenticationException();
         }
