+++ src/main/java/nextstep/auth/authentication/UserDetailsService.java
@@ -1,0 +1,7 @@
+package nextstep.auth.authentication;
+
+import nextstep.auth.User;
+
+public interface UserDetailsService {
+    User loadUserByUsername(String email);
+}
