+++ src/main/java/nextstep/auth/authentication/after/TokenAfterAuthentication.java
@@ -1,4 +1,4 @@
-package nextstep.auth.authentication;
+package nextstep.auth.authentication.after;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import nextstep.auth.context.Authentication;
