+++ src/test/java/nextstep/subway/unit/AuthenticationInterceptorTest.java
@@ -2,6 +2,12 @@ package nextstep.subway.unit;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import nextstep.auth.authentication.*;
+import nextstep.auth.authentication.after.AfterAuthentication;
+import nextstep.auth.authentication.after.SessionAfterAuthentication;
+import nextstep.auth.authentication.after.TokenAfterAuthentication;
+import nextstep.auth.authentication.converter.AuthenticationConverter;
+import nextstep.auth.authentication.converter.SessionAuthenticationConverter;
+import nextstep.auth.authentication.converter.TokenAuthenticationConverter;
 import nextstep.auth.context.Authentication;
 import nextstep.auth.context.SecurityContext;
 import nextstep.auth.token.JwtTokenProvider;
